services/doc-worker/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
python-dotenv
PyMuPDF
pytesseract
Pillow
opencv-python
sentence-transformers
numpy
requests
rich
python-multipart
